commit source

+++ CourseOnline/Models/GradeModels.cs
@@ -9,12 +9,12 @@ namespace CourseOnline.Models
     {
         public string user_fullname { get; set; }
         public string user_email { get; set; }
-        public int Test1 { get; set; }
-        public int Test2 { get; set; }
-        public int Test3 { get; set; }
-        public int Test4 { get; set; }
-        public int Test5 { get; set; }
-        public int Test6 { get; set; }
+        public double Test1 { get; set; }
+        public double Test2 { get; set; }
+        public double Test3 { get; set; }
+        public double Test4 { get; set; }
+        public double Test5 { get; set; }
+        public double Test6 { get; set; }
 
         public int coursework_weight { get; set; }
         public float Total { get; set; }
